android.webkit: Add scheme to test URI
The string "random" is no longer a valid historyUrl to pass to
WebView.loadDataWithBaseUrl().  The framework now validates it such
that the URL must have a scheme (any string before a colon will do).
I can't find a formal specification of this behavior, but intuition
says that this seems more likely to be a bug in the test than the
framework.

It's an upstream bug too, see:
https://code.google.com/p/android/issues/detail?id=32847

Change-Id: I789de11918334d4751eb1baa866df478a8470986
For:AXIA-2510
Signed-off-by: bma <Bo.Ma@windriver.com>